requirements.txt
python-dotenv
google-generativeai
SpeechRecognition
pyaudio #for Windows/Linux
#run brew install portaudio for Mac
